Move Engine::Veritas into {Engine, Relation::Graph::{Node, Edge}}
