--- lib/data_mapper/engine/veritas.rb
@@ -1,7 +1,0 @@
-require 'veritas'
-require 'veritas-do-adapter'
-
-require 'data_mapper/engine'
-require 'data_mapper/engine/veritas/engine'
-require 'data_mapper/engine/veritas/node'
-require 'data_mapper/engine/veritas/edge'
