--- spec/support/test_engine.rb
@@ -1,5 +1,0 @@
-class TestEngine < DataMapper::Engine::Veritas::Engine
-  def initialize(uri)
-    @relations = DataMapper::Relation::Graph.new(self)
-  end
-end
