+++ spec/unit/data_mapper/engine/gateway_relation_spec.rb
@@ -3,8 +3,11 @@ require 'spec_helper'
 describe Engine, '#gateway_relation' do
   subject { object.gateway_relation(relation) }
 
-  let(:object)   { subclass.new }
-  let(:relation) { mock('relation') }
+  let(:object) { described_class.new('postgres://localhost/test') }
 
-  it { should be(relation) }
+  let(:name)     { :users }
+  let(:header)   { [[:id, Integer], [:name, String]] }
+  let(:relation) { object.base_relation(name, header) }
+
+  it { should be_instance_of(Veritas::Relation::Gateway) }
 end
