--- spec/unit/data_mapper/engine/initialize_spec.rb
@@ -1,7 +1,0 @@
-require 'spec_helper'
-
-describe Engine, '#initialize' do
-  subject { object.new }
-
-  it_should_behave_like 'an abstract type'
-end
