--- spec/unit/data_mapper/engine/veritas/engine/gateway_relation_spec.rb
@@ -1,13 +1,0 @@
-require 'spec_helper'
-
-describe Engine::Veritas::Engine, '#gateway_relation' do
-  subject { object.gateway_relation(relation) }
-
-  let(:object) { described_class.new('postgres://localhost/test') }
-
-  let(:name)     { :users }
-  let(:header)   { [[:id, Integer], [:name, String]] }
-  let(:relation) { object.base_relation(name, header) }
-
-  it { should be_instance_of(Veritas::Relation::Gateway) }
-end
