--- spec/unit/data_mapper/engine/veritas/engine/relation_edge_class_spec.rb
@@ -1,9 +1,0 @@
-require 'spec_helper'
-
-describe Engine::Veritas::Engine, '#relation_edge_class' do
-  subject { object.relation_edge_class }
-
-  let(:object) { described_class.new('postgres://localhost/test') }
-
-  it { should be(Engine::Veritas::Edge) }
-end
