--- spec/unit/data_mapper/engine/veritas/node/each_spec.rb
@@ -1,22 +1,0 @@
-require 'spec_helper'
-
-describe Engine::Veritas::Node, '#each' do
-  let(:object)   { described_class.new(:users, relation) }
-  let(:relation) { mock('relation') }
-  let(:block)    { Proc.new {} }
-
-  context "with a block" do
-    subject { object.each(&block) }
-
-    it "passes the block to veritas relation" do
-      relation.should_receive(:each).with(&block)
-      subject
-    end
-  end
-
-  context "without a block" do
-    subject { object.each }
-
-    it { should be_instance_of(Enumerator) }
-  end
-end
