+++ spec/unit/data_mapper/relation/graph/node/each_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 
 describe Relation::Graph::Node, '#each' do
-  let(:object)   { subclass.new(name, relation) }
+  let(:object)   { described_class.new(name, relation) }
   let(:name)     { :users }
   let(:relation) { mock('relation') }
