+++ spec/unit/data_mapper/relation/graph/node/header_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe Engine::Veritas::Node, '#header' do
+describe Relation::Graph::Node, '#header' do
   subject { object.header }
 
   let(:object)   { described_class.new(:users, relation, header) }
